+++ pkg/runtime/collections/iterator.go
@@ -3,13 +3,10 @@ package collections
 import (
 	"context"
 	"github.com/MontFerret/ferret/pkg/runtime/core"
-	"github.com/MontFerret/ferret/pkg/runtime/values"
 )
 
 type (
 	Variables []string
-
-	DataSet map[string]core.Value
 
 	Iterator interface {
 		HasNext() bool
@@ -21,36 +18,6 @@ type (
 		Iterate(ctx context.Context, scope *core.Scope) (Iterator, error)
 	}
 )
-
-func (ds DataSet) Apply(scope *core.Scope, variables Variables) error {
-	if err := ValidateDataSet(ds, variables); err != nil {
-		return err
-	}
-
-	for _, variable := range variables {
-		if variable != "" {
-			value, found := ds[variable]
-
-			if !found {
-				return core.Errorf(core.ErrNotFound, "variable not found in a given data set: %s", variable)
-			}
-
-			scope.SetVariable(variable, value)
-		}
-	}
-
-	return nil
-}
-
-func (ds DataSet) Get(key string) core.Value {
-	val, found := ds[key]
-
-	if found {
-		return val
-	}
-
-	return values.None
-}
 
 func ToSlice(iterator Iterator) ([]DataSet, error) {
 	res := make([]DataSet, 0, 10)
